This fixes an issue with the handling of multi-line strings in cases where a block of text that normally would be interpreted as a multi-line string also contained a "special" line (such as a scope "{}" or array "[]").
In that case, the block should no longer be treated as a multi-line
value, but the parser did so.

This update fixes this by flushing the multi-line buffer when a scope or
array is encountered.

It also fixes an inconsistency with how newlines were treated. Newlines
are now included as part of every regex in the parsing loop, to ensure
consistent inclusion or disclusion of them in resulting multi-line
values.